Add warning for output dir and running process
In order to avoid outputing into existed directory and running
two process at the same time. We check the following two
conditions.
Check if output dir already existed in each analysis dialog when
we click ok.
Check if there is already a running process in AnalysisManager
after we click ok. If there is a ruuning process, then we send
signal to MainWindow to show warning message.