Fixed bug of setting file initialization
The setting file 'Setting.txt' will be created while we run the program at the first time. There is a typo 'Setting.pro' in function load. It creates Setting.txt instead of Setting.pro. Hence, there is an error when we run TFML at the first time, because there is no file named Setting.pro.